Rename 'enum' package to 'enumtool'
'enum' is a reserved keyword and it cannot be used as a package name.
