+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/DevToolsViewRegistry.kt
@@ -2,7 +2,7 @@ package com.maximbircu.devtools.android
 
 import android.content.Context
 import com.maximbircu.devtools.android.presentation.tool.DevToolLayout
-import com.maximbircu.devtools.android.presentation.tools.enum.EnumToolLayout
+import com.maximbircu.devtools.android.presentation.tools.enumtool.EnumToolLayout
 import com.maximbircu.devtools.android.presentation.tools.group.GroupToolLayout
 import com.maximbircu.devtools.android.presentation.tools.text.TextToolLayout
 import com.maximbircu.devtools.android.presentation.tools.time.TimeToolLayout
