+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/presentation/tools/enumtool/selectors/chips/EnumToolChipsOptionSelectorLayout.kt
@@ -1,4 +1,4 @@
-package com.maximbircu.devtools.android.presentation.tools.enum.selectors.chips
+package com.maximbircu.devtools.android.presentation.tools.enumtool.selectors.chips
 
 import android.annotation.SuppressLint
 import android.content.Context
