+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/presentation/tools/enumtool/selectors/dialog/EnumToolOptionSelectorDialog.kt
@@ -1,4 +1,4 @@
-package com.maximbircu.devtools.android.presentation.tools.enum.selectors.dialog
+package com.maximbircu.devtools.android.presentation.tools.enumtool.selectors.dialog
 
 import android.content.Context
 import androidx.recyclerview.widget.LinearLayoutManager
