+++ devtools/android/src/main/kotlin/com/maximbircu/devtools/android/presentation/tools/enumtool/selectors/dialog/EnumToolOptionsAdapterDelegate.kt
@@ -1,4 +1,4 @@
-package com.maximbircu.devtools.android.presentation.tools.enum.selectors.dialog
+package com.maximbircu.devtools.android.presentation.tools.enumtool.selectors.dialog
 
 interface EnumToolOptionsAdapterDelegate {
     fun getItemCount(): Int
